Extract Error and Result types for the Koto crate
