+++ CHANGELOG.md
@@ -121,6 +121,7 @@ The Koto project adheres to
   intermediate steps.
 - Implementing `KotoFile` has been made easier, with the `Display + Debug`
   constraint replaced with a required `id()` function.
+- `KotoError` and `KotoResult` are now `koto::Error` and `koto::Result`.
 
 ### Removed
