+++ core/koto/src/prelude.rs
@@ -1,5 +1,5 @@
 //! A collection of useful items to make it easier to work with `koto`
 
-pub use crate::{Koto, KotoError, KotoSettings};
+pub use crate::{Koto, KotoSettings};
 pub use koto_bytecode::{Chunk, Loader, LoaderError};
 pub use koto_runtime::prelude::*;
